refactor profile fragment.
