+++ app/src/main/java/com/linda/gourmetdiary/profile/ProfileFragment.kt
@@ -14,6 +14,7 @@ import com.github.mikephil.charting.data.*
 import com.github.mikephil.charting.formatter.IAxisValueFormatter
 import com.github.mikephil.charting.formatter.IndexAxisValueFormatter
 import com.linda.gourmetdiary.R
+import com.linda.gourmetdiary.data.Diaries4Day
 import com.linda.gourmetdiary.ext.getVmFactory
 import com.linda.gourmetdiary.databinding.ProfileFragmentBinding
 import com.linda.gourmetdiary.util.TimeConverters
@@ -25,17 +26,16 @@ class ProfileFragment : Fragment() {
 
     val viewModel by viewModels<ProfileViewModel> { getVmFactory() }
 
+    private lateinit var binding:ProfileFragmentBinding
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val binding = ProfileFragmentBinding.inflate(inflater, container, false)
+        binding = ProfileFragmentBinding.inflate(inflater, container, false)
 
         binding.lifecycleOwner = viewLifecycleOwner
         binding.viewModel = viewModel
-
-        val entries: MutableList<BarEntry> = ArrayList()
-        val label: MutableList<String> = mutableListOf()
 
         viewModel.diaryList.observe(viewLifecycleOwner, Observer { diary ->
             diary?.let {
@@ -45,65 +45,44 @@ class ProfileFragment : Fragment() {
 
         viewModel.diary4Day.observe(viewLifecycleOwner, Observer {
             it?.let { diaries ->
-
-                val xTitle = mutableListOf<String>()
-
-                for ((index, diary) in diaries.withIndex()) {
-
-                    val title = TimeConverters.timestampToDate(diary.dayTitle!!, Locale.TAIWAN)
-
-                    xTitle.add(title)
-                    label.add(TimeConverters.timestampToDate(diary.dayTitle!!, Locale.TAIWAN))
-
-                    var totalPriceForDay = 0
-                    diary.diaries.forEach { item ->
-                        totalPriceForDay += (item.food?.price?.toInt() ?: 0)
-                    }
-                    Log.d("Wayne","title=$title")
-                    Log.d("Wayne","totalPriceForDay=$totalPriceForDay}")
-                    Log.i("Wayne","----------------------------------------")
-
-                    entries.add(BarEntry(index.toFloat(), totalPriceForDay.toFloat()))
-                }
-                Log.d("xTitle","xTitle = $xTitle ; size = ${xTitle.size}")
-
-                /** 分隔線 **/
-
-                val dataSet = BarDataSet(entries, "每日花費")
-                dataSet.color = ContextCompat.getColor(requireContext(), R.color.colorAccent)
-                dataSet.valueTextColor = ContextCompat.getColor(requireContext(), R.color.kachi)
-
-                // Controlling X axis
-                val xAxis = binding.chart.xAxis
-                // Set the xAxis position to bottom. Default is top
-                xAxis.position = XAxis.XAxisPosition.BOTTOM
-                //Customizing x axis value
-                val months = arrayOf("M", "T", "W", "T", "F", "S", "S", "A", "A", "A")
-                val formatter = IAxisValueFormatter { value, axis -> months[value.toInt()] }
-                xAxis.granularity = 2f // minimum axis-step (interval) is 1
-
-                // Controlling right side of y axis
-                val yAxisRight = binding.chart.axisRight
-                yAxisRight.isEnabled = false
-
-                // Controlling left side of y axis
-                val yAxisLeft = binding.chart.axisLeft
-                yAxisLeft.granularity = 1f
-                // Setting Data
-                val data = BarData(dataSet)
-                binding.chart.data = data
-                binding.chart.invalidate()
-                binding.chart.xAxis.apply {
-                    valueFormatter = IndexAxisValueFormatter(label)
-                    labelCount = xTitle.size
-                    setDrawLabels(true)
-                    setDrawGridLines(false)
-                }
-                binding.chart.description.text = ""
-                binding.chart.setDrawGridBackground(false)
+                viewModel.assignChartData(diaries)
+                setBarChart()
             }
         })
 
         return binding.root
+    }
+
+    private fun setBarChart(){
+        val dataSet = BarDataSet(viewModel.entries, getString(R.string.daily_cost))
+        dataSet.color = ContextCompat.getColor(requireContext(), R.color.colorAccent)
+        dataSet.valueTextColor = ContextCompat.getColor(requireContext(), R.color.kachi)
+
+        // Controlling X axis
+        val xAxis = binding.chart.xAxis
+        // Set the xAxis position to bottom. Default is top
+        xAxis.position = XAxis.XAxisPosition.BOTTOM
+        //Customizing x axis value
+        xAxis.granularity = 2f // minimum axis-step (interval) is 1
+
+        // Controlling right side of y axis
+        val yAxisRight = binding.chart.axisRight
+        yAxisRight.isEnabled = false
+
+        // Controlling left side of y axis
+        val yAxisLeft = binding.chart.axisLeft
+        yAxisLeft.granularity = 1f
+        // Setting Data
+        val data = BarData(dataSet)
+        binding.chart.data = data
+        binding.chart.invalidate()
+        binding.chart.xAxis.apply {
+            valueFormatter = IndexAxisValueFormatter(viewModel.label)
+            labelCount = viewModel.xTitle.size
+            setDrawLabels(true)
+            setDrawGridLines(false)
+        }
+        binding.chart.description.text = ""
+        binding.chart.setDrawGridBackground(false)
     }
 }
