+++ app/src/main/res/values/strings.xml
@@ -67,6 +67,7 @@
     <string name="weekly_healthy">本週健康度：</string>
     <string name="search_template">尋找範本…</string>
     <string name="colon_text">：</string>
+    <string name="daily_cost">每日花費</string>
 
     <string name="can_booking">可訂位</string>
     <string name="cannot_booking">不可訂位</string>
